Use CHORDAL_HOLD
This makes is easier to quickly type things that have 'z' or '/'
followed by another key without invoking the 'hold' action (Ctrl).   If
the second key is on the same hand then the 'tap' action is used even if
the 'hold' action would have been used according to the PERMISSIVE_HOLD
or HOLD_ON_OTHER_KEY_PRESS setting.